Initial Commit
This is very loosely based on the current twitch pubsub overlay panel
template. This sets things up so that the panel controls will follow an
appropriate state, to whit:

  - If there is no video name (or it's empty), nothing will work
  - the previous and next buttons will change the scene number
  - the previous and next buttons will disable if the scene is invalid
  - the previous button will disable if the scene is <= 1
  - Pressing enter or clicking a button transmits a new template

Whenever a new filename template is submitted, we disable the form
controls, send up the new template, then ask what the template is and
use that to re-enable and update the form.